svelte: Add search query input to repo pages
This PR adds the search query input to the repo page header. I used a
similar solution as GitHub, which renders a button that looks a bit like
an input and opens a dialog with the real input.
The input is prepopulated with a `repo:` filter for the current input.
It has a tab/focus trap and closes on escape.

Before settling on this solution I tried various other approaches:

- Directly render the input in the header and make it expand on focus.
  That caused strange text selection because mousdown would expand the
  input and mouseup would happen at a different place in the input,
  causing text selection.
- Using the `<dialog>` element. It has some default styles/behavior that
  interfered with what I wanted to do.
- I wanted to use the newly added `openFocus` option in `createDialog`
  to disable melt's autofocus behavior, but that didn't work.

Happy to iterate on the design of the button, or on the behavior in
general.

I decided against introducing keyboard shortcuts in this PR, I think
that warrants it's own PR.